Several changes.
Storing terminalSequenceLengths as bytes instead of ints in HierarchicalPhrases to save memory.
Deleted unused class.
Added more detailed logging in ExtractRules to help trace memory usage with respect to the Cache in SuffixArray.


git-svn-id: https://joshua.svn.sf.net/svnroot/joshua/trunk@532 0ae5e6b2-d358-4f09-a895-f82f13dd62a4

+++ src/joshua/sarray/HierarchicalPhrases.java
@@ -47,8 +47,15 @@ public class HierarchicalPhrases {
 	/** 
 	 * Represents the length of each 
 	 * contiguous sequence of terminals in the pattern. 
+	 * <p>
+	 * To save memory, this information is stored as 
+	 * bytes instead of integers. 
+	 * 
+	 * This means that the maximum value that can be
+	 * stored here is 127. This should not be a problem
+	 * unless a very large value is used for maximum phrase length.
 	 */
-	private final int[] terminalSequenceLengths;
+	private final byte[] terminalSequenceLengths;
 	
 	
 	/** Number of hierarchical phrases represented by this object. */
